fix(rubocop): extract search instrumentation event modules
First slice of warehouse 9af971264 design: move QueryEvents, ApiEvents,
ResultEvents, and EvaluationEvents into nested modules under
Search::Instrumentation. Preserve AiUsage payload behaviour on api_call.
Core/payload/summary helpers remain for a follow-up slice under 800 lines.

Issue: No ticket/issue
