+++ app/lib/search/instrumentation.rb
@@ -4,6 +4,11 @@ module Search
   module Instrumentation
   module_function
 
+    extend QueryEvents
+    extend ApiEvents
+    extend ResultEvents
+    extend EvaluationEvents
+
     ERROR_MESSAGE_MAX_LENGTH = 500
     MAX_LOGGED_RESULTS = 50
     EVALUATION_TRACE_VERSION = 'classification_evaluation_trace.v1'.freeze
@@ -29,260 +34,6 @@ module Search
     rescue StandardError => e
       search_failed(request_id:, error_type: e.class.name, error_message: e.message, search_type:)
       raise
-    end
-
-    def query_expanded(request_id:, original_query:)
-      start_time = Process.clock_gettime(Process::CLOCK_MONOTONIC)
-      result = yield
-      duration = Process.clock_gettime(Process::CLOCK_MONOTONIC) - start_time
-
-      instrument(
-        'query_expanded',
-        request_id:,
-        search_type: 'interactive',
-        original_query:,
-        expanded_query: result.expanded_query,
-        reason: result.reason,
-        duration_ms: (duration * 1000).round(2),
-      )
-
-      result
-    end
-
-    def query_refined(request_id:, original_query:, refined_query:, answer_count:, base_query: original_query, effective_query: refined_query, added_answers: [], iteration: nil)
-      result = yield
-      instrument(
-        'query_refined',
-        request_id:,
-        search_type: 'interactive',
-        base_query:,
-        original_query:,
-        refined_query:,
-        effective_query:,
-        answer_count:,
-        added_answers:,
-        iteration:,
-      )
-      result
-    end
-
-    def query_expansion_decided(request_id:, query:, expand:, reason:, result_count:, max_score:)
-      instrument('query_expansion_decided', request_id:, search_type: 'interactive', query:, expand:, reason:, result_count:, max_score:)
-    end
-
-    def api_call(request_id:, model:, attempt_number:, iteration: nil, effective_query: nil, operation: 'interactive_search', emit_search_failed: true)
-      start_time = Process.clock_gettime(Process::CLOCK_MONOTONIC)
-      result = yield
-      duration = Process.clock_gettime(Process::CLOCK_MONOTONIC) - start_time
-
-      instrument(
-        'api_call_completed',
-        {
-          request_id:,
-          search_type: 'interactive',
-          model:,
-          duration_ms: (duration * 1000).round(2),
-          response_type: determine_response_type(result),
-          attempt_number:,
-          iteration:,
-          effective_query:,
-          operation:,
-          event_kind: operation,
-        }.merge(error_payload_for_result(result)).merge(AiUsage.payload_from(result)),
-      )
-
-      result
-    rescue StandardError => e
-      duration = Process.clock_gettime(Process::CLOCK_MONOTONIC) - start_time
-      instrument(
-        'api_call_completed',
-        {
-          request_id:,
-          search_type: 'interactive',
-          model:,
-          duration_ms: (duration * 1000).round(2),
-          response_type: 'error',
-          attempt_number:,
-          iteration:,
-          effective_query:,
-          operation:,
-          event_kind: operation,
-        }.merge(truncate_error_payload(AiUsage.safe_error_message(e))).merge(AiUsage.payload_from_error(e)),
-      )
-      search_failed(request_id:, error_type: e.class.name, error_message: AiUsage.safe_error_message(e), search_type: 'interactive') if emit_search_failed
-      raise
-    end
-
-    def exact_match_selected(request_id:, search_type:, query:, match_source:, matched_value:, result:)
-      result_fields = result_summary(result)
-      instrument(
-        'exact_match_selected',
-        {
-          request_id:,
-          search_type:,
-          query:,
-          match_source:,
-          matched_value:,
-          target_type: result_fields[:goods_nomenclature_class],
-          target_id: result_fields[:target_id],
-          target_endpoint: result_fields[:target_endpoint],
-          goods_nomenclature_item_id: result_fields[:goods_nomenclature_item_id],
-          goods_nomenclature_sid: result_fields[:goods_nomenclature_sid],
-          details: result_fields,
-        },
-      )
-    end
-
-    def fuzzy_results_returned(request_id:, query:, results:)
-      instrument(
-        'fuzzy_results_returned',
-        request_id:,
-        search_type: 'classic',
-        query:,
-        result_count: nested_result_count(results),
-        details: summarize_classic_fuzzy_results(results),
-      )
-    end
-
-    def interactive_configuration_used(request_id:, query:, configuration:)
-      instrument(
-        'interactive_configuration_used',
-        request_id:,
-        search_type: 'interactive',
-        query:,
-        details: configuration,
-      )
-    end
-
-    def retrieval_results_returned(request_id:, query:, search_type:, retrieval_method:, stage:, results:, effective_query: nil, leg: nil, iteration: nil)
-      instrument(
-        'retrieval_results_returned',
-        request_id:,
-        search_type:,
-        query:,
-        effective_query:,
-        retrieval_method:,
-        stage:,
-        leg:,
-        iteration:,
-        result_count: Array(results).size,
-        details: { results: summarize_results(results) },
-      )
-    end
-
-    def question_returned(request_id:, question_count:, attempt_number:, iteration: nil, effective_query: nil, questions: nil)
-      payload = { request_id:, search_type: 'interactive', question_count:, attempt_number:, iteration:, effective_query: }
-      payload[:details] = { questions: questions } if questions
-      instrument('question_returned', payload)
-    end
-
-    def answer_returned(request_id:, answer_count:, confidence_levels:, attempt_number:, iteration: nil, effective_query: nil, answers: nil)
-      payload = { request_id:, search_type: 'interactive', answer_count:, confidence_levels:, attempt_number:, iteration:, effective_query: }
-      payload[:details] = { answers: answers } if answers
-      instrument('answer_returned', payload)
-    end
-
-    def evaluation_trace_returned(request_id:, query:, effective_query:, iteration:, answer_count:, retrieval_method:, results_type:, candidates:, final_result_type:, ranked_answers:, questions:, error_message:, ranking_source:, model:, result_limit:)
-      candidate_summaries = summarize_results(candidates)
-      ranked_answer_summaries = summarize_ranked_answers(ranked_answers)
-      question_summaries = summarize_questions(questions)
-
-      instrument(
-        'evaluation_trace_returned',
-        request_id:,
-        search_type: 'interactive',
-        trace_version: EVALUATION_TRACE_VERSION,
-        query:,
-        effective_query:,
-        iteration:,
-        answer_count:,
-        retrieval_method:,
-        results_type:,
-        candidate_count: Array(candidates).size,
-        logged_candidate_count: candidate_summaries.size,
-        candidates_truncated: candidate_summaries.size < Array(candidates).size,
-        final_result_type:,
-        ranked_answer_count: Array(ranked_answers).size,
-        logged_ranked_answer_count: ranked_answer_summaries.size,
-        ranked_answers_truncated: ranked_answer_summaries.size < Array(ranked_answers).size,
-        question_count: Array(questions).size,
-        logged_question_count: question_summaries.size,
-        questions_truncated: question_summaries.size < Array(questions).size,
-        confidence_levels: confidence_levels_for(ranked_answer_summaries),
-        ranking_source:,
-        model:,
-        result_limit:,
-        details: {
-          candidates: candidate_summaries,
-          ranked_answers: ranked_answer_summaries,
-          questions: question_summaries,
-        }.compact_blank,
-        **truncate_error_payload(error_message),
-      )
-    end
-
-    def duplicate_question_guard_checked(request_id:, attempt_number:, allowed:, duplicate:, suspicious:, signals:, reason:, iteration: nil, effective_query: nil, duplicate_of_question: nil, duplicate_of_answer: nil)
-      reason_payload = truncate_reason_payload(reason)
-
-      instrument(
-        'duplicate_question_guard_checked',
-        request_id:,
-        search_type: 'interactive',
-        attempt_number:,
-        iteration:,
-        effective_query:,
-        allowed:,
-        duplicate:,
-        suspicious:,
-        signals:,
-        reason: reason_payload[:reason],
-        reason_truncated: reason_payload[:reason_truncated],
-        duplicate_of_question:,
-        duplicate_of_answer:,
-      )
-    end
-
-    def search_completed(request_id:, search_type:, total_duration_ms:, result_count:, query: nil, total_attempts: nil, total_questions: nil, final_result_type: nil, results_type: nil, max_score: nil, error_message: nil, description_intercept: nil)
-      payload = {
-        request_id:,
-        query:,
-        search_type:,
-        total_attempts:,
-        total_questions:,
-        final_result_type:,
-        total_duration_ms:,
-        result_count:,
-      }
-      payload[:results_type] = results_type if results_type
-      payload[:max_score] = max_score if max_score
-      payload.merge!(description_intercept_payload(description_intercept, prefix: :description_intercept))
-      payload.merge!(truncate_error_payload(error_message))
-      instrument('search_completed', payload)
-    end
-
-    def description_intercept_checked(request_id:, query:, description_intercept:)
-      instrument(
-        'description_intercept_checked',
-        { request_id:, search_type: 'interactive', query: }.merge(description_intercept_payload(description_intercept)),
-      )
-    end
-
-    def retrieval_leg_completed(request_id:, leg:, duration_ms:, result_count:, status:, error_message: nil)
-      instrument(
-        'retrieval_leg_completed',
-        {
-          request_id:,
-          search_type: 'interactive',
-          leg:,
-          duration_ms:,
-          result_count:,
-          status:,
-        }.merge(truncate_error_payload(error_message)),
-      )
-    end
-
-    def result_selected(request_id:, goods_nomenclature_item_id:, goods_nomenclature_class:)
-      instrument('result_selected', request_id:, goods_nomenclature_item_id:, goods_nomenclature_class:)
     end
 
     def search_failed(request_id:, error_type:, error_message:, search_type:)
@@ -315,33 +66,6 @@ module Search
       return payload unless prefix
 
       payload.transform_keys { |key| [prefix, key].join('_').to_sym }
-    end
-
-    def determine_response_type(result)
-      return 'unknown' if result.nil?
-
-      parsed = parse_result(result)
-      return 'error' if parsed.is_a?(Hash) && parsed['error'].present?
-      return 'answers' if parsed.is_a?(Hash) && parsed['answers'].present?
-      return 'questions' if parsed.is_a?(Hash) && parsed['questions'].is_a?(Array) && parsed['questions'].any?
-      return 'duplicate_validation' if parsed.is_a?(Hash) && [true, false].include?(parsed['duplicate'])
-
-      'unknown'
-    rescue StandardError
-      'unknown'
-    end
-
-    def error_payload_for_result(result)
-      parsed = parse_result(result)
-      return {} unless parsed.is_a?(Hash) && parsed['error'].present?
-
-      truncate_error_payload(parsed['error'])
-    rescue StandardError
-      {}
-    end
-
-    def parse_result(result)
-      result.is_a?(String) ? ExtractBottomJson.call(result) : result
     end
 
     def truncate_error_payload(error_message)
